restructured the client side code tto handle react-router-dom in a cleaner organizational way. This was accomplished By moving rendering to only components and only having containers be responsible for structuring dispatch methods and bringing in state from the store.
