+++ client/containers/MainContainer.jsx
@@ -6,81 +6,31 @@
 /* eslint-disable react */
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import Navigation from './NavigationContainer.jsx';
-import Cards from './CardsContainer.jsx';
-import types from '../constants/actionTypes';
-import * as actions from '../actions/actions';
-import ItemForm from '../components/ItemForm.jsx';
+import { Route, Switch, withRouter } from 'react-router-dom';
 import MessagesContainer from './MessagesContainer.jsx';
-
-const mapStateToProps = store => ({
-  cards: store.cards
-});
-
-// need to add all our action creators here
-const mapDispatchToProps = dispatch => ({
-  fetchAllItems: () => {
-    dispatch(actions.fetchItemsData());
-  },
-  fetchSearchedItems: value => {
-    dispatch(actions.fetchSearchedItems(value));
-  },
-  searchBoxChange: value => {
-    dispatch(actions.searchValueChange(value));
-  },
-  fetchCategory: value => {
-    dispatch(actions.fetchCategoryItems(value));
-  },
-  addItem: event => {
-    event.preventDefault();
-    const addItemForm = document.getElementById('addItemForm');
-    const formData = {
-      user_id: 1,
-      item_name: addItemForm.elements.name.value,
-      item_details: addItemForm.elements.desc.value,
-      price: addItemForm.elements.price.value,
-      photo: addItemForm.elements.url.value,
-    }
-    dispatch(actions.addItem(formData))
-  }
-});
+import CardsContainer from './CardsContainer.jsx';
 
 class MainContainer extends Component {
   constructor(props) {
     super(props);
   }
 
-  componentWillMount() {
-    this.props.fetchAllItems();
-  }
-
   render() {
     return (
       <div>
-        <div id="navdiv">
-          <Navigation
-            fetchSearchedItems={this.props.fetchSearchedItems}
-            fetchCategory={this.props.fetchCategory}
-            searchValue={this.props.cards.searchBoxValue}
-            searchBoxChange={this.props.searchBoxChange}
+        <Switch>
+          <Route 
+            exact path="/"
+            component={CardsContainer}
           />
-        </div>
-        {/* <div id="cardsdiv">
-          <Cards
-            items={this.props.cards.items}
-            fetchFlag={this.props.cards.fetching}
-            loading={this.props}
+          <Route 
+            exact path="/messages"
+            component={MessagesContainer}
           />
-  
-        </div> */}
-        <MessagesContainer />
-        {/* <ItemForm addItem={this.props.addItem} /> */}
+        </Switch>
       </div>
     );
   }
 }
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(MainContainer);
+export default withRouter(connect(null, null)(MainContainer));
